refactor:componentsフォルダの位置を変更、不要なファイルを削除
