+++ app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import Footer from "./components/organisms/footer";
-import ClientSideComponent from './components/ClientSideComponent';
+import Footer from "../components/organisms/footer";
+import ClientSideComponent from '../components/ClientSideComponent';
 
 const inter = Inter({ subsets: ["latin"] });
