+++ components/ClientSideComponent.tsx
