+++ components/organisms/petStatusModal.tsx
@@ -1,4 +1,3 @@
-
 import React from 'react';
 import { Box, Typography, Button, Modal } from '@mui/material';
